// src/components/types.ts
export interface coinTypes {
	id: number;
	market_cap_rank: number;
	market_cap: number;
	image: string;
	current_price: number;
	price_change_percentage_1h_in_currency: number;
	price_change_percentage_24h: number;
	price_change_percentage_24h_in_currency: number;
	price_change_percentage_7d_in_currency: number;
}